Revert "Feature/swift package manager"
