--- Package.resolved
@@ -1,16 +1,0 @@
-{
-  "object": {
-    "pins": [
-      {
-        "package": "SwiftLazy",
-        "repositoryURL": "https://github.com/Nekitosss/SwiftLazy.git",
-        "state": {
-          "branch": null,
-          "revision": "19a75bdec163fb6c3cbee0aced624f98df50a7c0",
-          "version": "1.2.1"
-        }
-      }
-    ]
-  },
-  "version": 1
-}
